fix(calling): extended eventing class on interface ILine
