+++ packages/calling/src/CallingClient/line/index.ts
@@ -9,7 +9,7 @@ import {
   RegistrationStatus,
   ServiceIndicator,
 } from '../../common/types';
-import {ILine, LINE_EVENTS, LineEventTypes} from './types';
+import {ILine, LINE_EVENTS} from './types';
 import {LINE_FILE, VALID_PHONE} from '../constants';
 import log from '../../Logger';
 import {IRegistration} from '../registration/types';
@@ -21,7 +21,7 @@ import {LineError} from '../../Errors/catalog/LineError';
 import {LOGGER} from '../../Logger/types';
 import {validateServiceData} from '../../common';
 import SDKConnector from '../../SDKConnector';
-import {LINE_EVENT_KEYS} from '../../Events/types';
+import {LINE_EVENT_KEYS, LineEventTypes} from '../../Events/types';
 import {ICall, ICallManager} from '../calling/types';
 import {getCallManager} from '../calling/callManager';
 import {ERROR_TYPE} from '../../Errors/types';
@@ -43,13 +43,11 @@ export default class Line extends Eventing<LineEventTypes> implements ILine {
 
   public mobiusDeviceId?: string;
 
-  private mobiusUri?: string;
-
   public phoneNumber?: string;
 
   public extension?: string;
 
-  public sipAddresses?: string[];
+  public sipAddresses: string[] = [];
 
   public voicemail?: string;
