+++ packages/calling/src/index.ts
@@ -50,3 +50,4 @@ export {
 } from './common/types';
 export {CallError, LineError} from './Errors';
 export {ICall, TransferType} from './CallingClient/calling/types';
+export {LOGGER} from './Logger/types';
